requirements.txt
GPUtil==1.4.0
flask==2.3.3
google-cloud-pubsub==2.18.4
redis==5.0.0